An small mistake i did earlier whe i tried to update the README... still a wip anyways

+++ libhfs/os/unix.c
@@ -48,6 +48,7 @@ int fstat(int, struct stat *);
 
 # include <errno.h>
 # include <sys/stat.h>
+# include <stdint.h>
 
 # include "libhfs.h"
 # include "os.h"
@@ -88,7 +89,7 @@ int os_open(void **priv, const char *path, int mode)
       (errno == EACCES || errno == EAGAIN))
     ERROR(EAGAIN, "unable to obtain lock for medium");
 
-  *priv = (void *) fd;
+  *priv = (void *) (intptr_t) fd;
 
   return 0;
 
@@ -105,9 +106,9 @@ fail:
  */
 int os_close(void **priv)
 {
-  int fd = (int) *priv;
+  int fd = (int) (intptr_t) *priv;
 
-  *priv = (void *) -1;
+  *priv = (void *) (intptr_t) -1;
 
   if (close(fd) == -1)
     ERROR(errno, "error closing medium");
@@ -124,7 +125,7 @@ fail:
  */
 int os_same(void **priv, const char *path)
 {
-  int fd = (int) *priv;
+  int fd = (int) (intptr_t) *priv;
   struct stat fdev, dev;
 
   if (fstat(fd, &fdev) == -1 ||
@@ -144,7 +145,7 @@ fail:
  */
 unsigned long os_seek(void **priv, unsigned long offset)
 {
-  int fd = (int) *priv;
+  int fd = (int) (intptr_t) *priv;
   off_t result;
 
   /* offset == -1 special; seek to last block of device */
@@ -169,7 +170,7 @@ fail:
  */
 unsigned long os_read(void **priv, void *buf, unsigned long len)
 {
-  int fd = (int) *priv;
+  int fd = (int) (intptr_t) *priv;
   ssize_t result;
 
   result = read(fd, buf, len << HFS_BLOCKSZ_BITS);
@@ -189,7 +190,7 @@ fail:
  */
 unsigned long os_write(void **priv, const void *buf, unsigned long len)
 {
-  int fd = (int) *priv;
+  int fd = (int) (intptr_t) *priv;
   ssize_t result;
 
   result = write(fd, buf, len << HFS_BLOCKSZ_BITS);
